+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 9.0.1 (2022-12-16)
+Features:
+- Implement Xendit metric
+- Fix get payment_method from URL query
+
 ## 9.0.0 (2022-10-26)
 Features:
 - Add new IDR payment: Atome